Add funcs and types in azure/fnNetworkIpAddresses.go
