+++ lib/wrapperFuncs.go
@@ -6,6 +6,8 @@ import (
 	"log"
 	"runtime"
 	"strconv"
+	"sync"
+	"sync/atomic"
 )
 
 // Simple wrapper to
@@ -20,7 +22,7 @@ func CheckFatalError(err error) {
 	}
 }
 
-func PrintSrcLoc(str string) {
+func PrintSrcLoc(str ...string) {
 	_, file, no, _ := runtime.Caller(1)
 	fmt.Println(file+":"+strconv.Itoa(no)+":0", str)
 }
@@ -38,4 +40,23 @@ func MarshalAndPrintJson(data any) {
 		log.Fatal(err)
 	}
 	fmt.Println(string(jsonStr))
+}
+
+type WaitGroupCount struct {
+	sync.WaitGroup
+	count int64
+}
+
+func (wg *WaitGroupCount) Add(delta int) {
+	atomic.AddInt64(&wg.count, int64(delta))
+	wg.WaitGroup.Add(delta)
+}
+
+func (wg *WaitGroupCount) Done() {
+	atomic.AddInt64(&wg.count, -1)
+	wg.WaitGroup.Done()
+}
+
+func (wg *WaitGroupCount) GetCount() int {
+	return int(atomic.LoadInt64(&wg.count))
 }
